+++ src/lib/components/carousel/index.vue
@@ -401,5 +401,3 @@ export default {
 };
 </script>
 
-<style lang="less" src="./index.less"></style>
-