+++ src/lib/components/empty/index.vue
@@ -68,5 +68,3 @@ export default {
     mounted : function () {}
 };
 </script>
-
-<style lang="less" src="./index.less"></style>