+++ src/lib/components/img/index.vue
@@ -29,5 +29,3 @@ export default {
     }
 };
 </script>
-
-<style lang="less" src="./index.less"></style>