+++ src/lib/components/message/index.vue
@@ -224,5 +224,3 @@ export default {
     }
 };
 </script>
-
-<style lang="less" src="./index.less"></style>