+++ src/lib/components/switch/index.vue
@@ -103,5 +103,3 @@ export default {
     mounted : function () {}
 };
 </script>
-
-<style lang="less" src="./index.less"></style>