+++ src/lib/components/video/index.vue
@@ -182,5 +182,3 @@ export default {
     }
 };
 </script>
-
-<style lang="less" src="./index.less"></style>