+++ CHANGELOG.md
@@ -1,5 +1,56 @@
 # CHANGELOG
 
+## 0.12.2
+
+`2019年2月22日`
+
+#### 新特性
+
+- 支持自定义主题色
+
+#### 改进
+
+- 现在`ui-dropdown`、`ui-tip`、`ui-popover`在鼠标右键点击时不会触发弹出框的显示及隐藏
+
+#### 修复
+
+- 修复`ui-dropdown`当`trigger`配置不为`click`时，点击展示下拉列表的逻辑问题
+- 修复`ui-table`文档中的一些问题
+- 修复`ui-calendar`文档中的一处问题及部分文档日期格式在`Safari`浏览器中错误的问题
+- 修复`ui-tip`组件当`trigger`配置设为`hover click`时的触发行为不正确的问题
+- 修复`ui-carousel`文档中的一处错误
+- 修复`ui-menu`在使用`side-collapse`配置及`block`样式时的交互样式错误问题
+
+#### 更多
+
+- 所有交互组件可兼容`Firefox`及`Safari`浏览器
+- 组件库中所有的`childs`字段均改为`children`
+
+__升级注意事项__
+
+- 所有组件配置中的`childs`字段已改为`children`，涉及：`ui-cascader`的`list`配置、`ui-menu`的`menu`配置、`ui-tree`的`tree`配置
+
+
+## 0.12.1
+
+`2019年2月19日`
+
+#### 新特性
+
+- 新增`ui-empty`空状态组件
+- `ui-transfer`新增`source-to-target-text`和`target-to-source-text`配置，可以自定义按钮文案
+- `ui-transfer`新增`source-footer`和`target-footer`两个`slot`，可以自定义两侧底部内容
+
+#### 修复
+
+- 修复`ui-tree`的`can-click`配置设为`false`时，子节点仍然可以点击的问题
+- 修复`ui-tree`的连接线(`cable`配置)的样式问题
+- 修复一些文档错误
+
+#### 更多
+
+- 新增图标( `mo-icon-tag-f` / `mo-icon-drawer-full` / `mo-icon-drawer-empty`)
+
 ## 0.12.0
 
 `2019年2月8日`